backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
alembic
pydantic
passlib[bcrypt]
python-jose[cryptography]
pytest
python-multipart
pydantic-settings
